更新 属性页信息显示 显示工作流内容

+++ ByteCode.cpp
@@ -20,13 +20,13 @@ int CByteCode::UTF_8ToUnicode(CString& sOut, char* szU8)
 
 
 	//预转换，得到所需空间的大小 
-	int wcsLen = ::MultiByteToWideChar(CP_UTF8, NULL, szU8, strlen(szU8), NULL, 0);
+	int wcsLen = ::MultiByteToWideChar(CP_UTF8, NULL, szU8, (int)strlen(szU8), NULL, 0);
 
 	//分配空间要给'\0'留个空间，MultiByteToWideChar不会给'\0'空间 
 	wchar_t* wszString = new wchar_t[wcsLen + 1];
 
 	//转换 
-	::MultiByteToWideChar(CP_UTF8, NULL, szU8, strlen(szU8), wszString, wcsLen);
+	::MultiByteToWideChar(CP_UTF8, NULL, szU8, (int)strlen(szU8), wszString, wcsLen);
 
 	//最后加上'\0' 
 	wszString[wcsLen] = '\0';
@@ -41,10 +41,10 @@ void CByteCode::UnicodeToUTF_8(std::string& pOut, CString& pText)
 	// unicode to UTF8 
 	wchar_t* wszString = pText.GetBuffer();
 	pText.ReleaseBuffer();
-	int u8Len = ::WideCharToMultiByte(CP_UTF8, NULL, wszString, wcslen(wszString), NULL, 0, NULL, NULL);
+	int u8Len = ::WideCharToMultiByte(CP_UTF8, NULL, wszString, (int)wcslen(wszString), NULL, 0, NULL, NULL);
 
 	char* szU8 = new char[u8Len + 1];
-	::WideCharToMultiByte(CP_UTF8, NULL, wszString, wcslen(wszString), szU8, u8Len, NULL, NULL);
+	::WideCharToMultiByte(CP_UTF8, NULL, wszString, (int)wcslen(wszString), szU8, u8Len, NULL, NULL);
 	szU8[u8Len] = '\0';
 	pOut = szU8;
